JPA Server side test fixes - part 1 (#2401)
* JPA Server side test fixes - part 1
This is first part of the fix for JPA test apps executed in JEE server environment.
Main changes there:
- eclipselink-testbuild-plugin upgrade to 1.1.2 with <libs> support
- (JpaEntityManager)entityManager -> entityManager.unwrap(JpaEntityManager) as server should provide proxy class
- em.getTransaction().begin() -> beginTransaction(em) to allow JPA test FW identify if test is executed in JSE or JEE environment and call proper code

+++ jpa/eclipselink.jpa.testapps/jpa.test.xml.only/jpa.test.xml.advanced/src/test/java/org/eclipse/persistence/testing/tests/jpa/xml/advanced/XmlAdvancedJunitTest.java
@@ -72,30 +72,33 @@ public class XmlAdvancedJunitTest extends JUnitTestCase {
     }*/
 
     public void testEL254937(){
-        EntityManager em = createEntityManager();
-        beginTransaction(em);
-        LargeProject lp1 = new LargeProject();
-        lp1.setName("one");
-        em.persist(lp1);
-        commitTransaction(em);
-        em = createEntityManager();
-        beginTransaction(em);
-        em.remove(em.find(LargeProject.class, lp1.getId()));
-        em.flush();
-        JpaEntityManager eclipselinkEm = (JpaEntityManager)em.getDelegate();
-        RepeatableWriteUnitOfWork uow =
-            (RepeatableWriteUnitOfWork)eclipselinkEm.getActiveSession();
-        //duplicate the beforeCompletion call
-        uow.issueSQLbeforeCompletion();
-        //commit the transaction
-        uow.setShouldTerminateTransaction(true);
-        uow.commitTransaction();
-        //duplicate the AfterCompletion call.  This should merge, removing the LargeProject from the shared cache
-        uow.mergeClonesAfterCompletion();
-        em = createEntityManager();
-        LargeProject cachedLargeProject = em.find(LargeProject.class, lp1.getId());
-        closeEntityManager(em);
-        assertNull("Entity removed during flush was not removed from the shared cache on commit", cachedLargeProject);
+        // Should not run in the server - bug 264589
+        if (! isOnServer()) {
+            EntityManager em = createEntityManager();
+            beginTransaction(em);
+            LargeProject lp1 = new LargeProject();
+            lp1.setName("one");
+            em.persist(lp1);
+            commitTransaction(em);
+            em = createEntityManager();
+            beginTransaction(em);
+            em.remove(em.find(LargeProject.class, lp1.getId()));
+            em.flush();
+            JpaEntityManager eclipselinkEm = (JpaEntityManager) em.getDelegate();
+            RepeatableWriteUnitOfWork uow =
+                    (RepeatableWriteUnitOfWork) eclipselinkEm.getActiveSession();
+            //duplicate the beforeCompletion call
+            uow.issueSQLbeforeCompletion();
+            //commit the transaction
+            uow.setShouldTerminateTransaction(true);
+            uow.commitTransaction();
+            //duplicate the AfterCompletion call.  This should merge, removing the LargeProject from the shared cache
+            uow.mergeClonesAfterCompletion();
+            em = createEntityManager();
+            LargeProject cachedLargeProject = em.find(LargeProject.class, lp1.getId());
+            closeEntityManager(em);
+            assertNull("Entity removed during flush was not removed from the shared cache on commit", cachedLargeProject);
+        }
     }
 
     public void testGF1894() {
